+++ skued/simulation/__init__.py
@@ -2,4 +2,4 @@
 
 from .powdersim import powdersim
 from .potential import electrostatic, pelectrostatic
-from .structure_factors import structure_factor, bounded_reflections, affe
+from .structure_factors import structure_factor, affe